add 3層構造にした
