+++ src/app/_features/OnboardingDialog/OnboardingDialog.tsx
@@ -1,29 +1,15 @@
 "use client";
 
-import { useGetRecommendedUsers } from "@/app/hooks/useGetRecommendedUsers";
-import { useState } from "react";
 import { OnboardingDialogView } from "./OnboardingDialogView";
+import { useOnboardingDialog } from "./useOnboardingDialog";
 
 type OnboardingProps = {
   onClose: () => void;
 };
 
 export function OnboardingDialog({ onClose }: OnboardingProps) {
-  const [currentStep, setCurrentStep] = useState(1);
-  const [name, setName] = useState("");
-  const [interests, setInterests] = useState<string[]>([]);
-
-  // フックを使用しておすすめのユーザーを取得
-  const { recommendedUsers = [], isLoading } = useGetRecommendedUsers();
-
-  const handleNext = () => {
-    if (currentStep < 3) {
-      setCurrentStep(currentStep + 1);
-    } else {
-      console.log("登録完了:", { name, interests });
-      onClose();
-    }
-  };
+  const { currentStep, name, setName, interests, setInterests, recommendedUsers, isLoading, handleNext } =
+    useOnboardingDialog(onClose);
 
   return (
     <OnboardingDialogView
